// src/app/api/model/flashcardResponse.ts
/**
 * Flashcards API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Flashcard } from './flashcard';


export interface FlashcardResponse { 
    message: string;
    cards: Array<Flashcard>;
}